Make VFS work and adapt code to use it
Resolve filenames recursively in Resolve().
Make VFS::FileExists() use Resolve().
Add PushState(), PopState() and Mount().

Remove path resolving code in paths.cpp, make code use the one in
vfs.cpp instead.

Also make the vfs resolver able to differentiate a resolve for a file
that is to be written to and a file that is to be read from,
and improve code in filesystem.cpp to support this in a better way.

Adapt code to be able to use VFS code properly, by looking in the right
directories.
Also changes to expected prefixes to files; for example textures will
have an implied prefix of /data/textures/ and so on, so data files had
to be modified to be less specific about file names.
Also put level scripts in 'mountable' directories; more stuff will be
moved there later on. A directory for a level is now mounted when it is
loaded, so files specific to a level can be kept in a separate
directory. Also, when the VFS system gains archive support, a full level
may be put in for example a tar archive or a zip file.

git-svn-id: https://bloopz.zapto.org/svn/nightfall/trunk@1332 d57568f3-fe0c-0410-b608-e86116ac1d00